% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_type_geometrie.R
\name{add_type_geometrie}
\alias{add_type_geometrie}
\title{Ajoute une colonne contenant le type de geometrie pour les observations}
\usage{
add_type_geometrie(oison_df)
}
\arguments{
\item{oison_df}{un dataframe de classe sf}
}
\value{
un dataframe avec une colonne `type_geometrie`
}
\description{
Ajoute une colonne contenant le type de geometrie pour les observations
}
\examples{
\dontrun{
oison_taxon <-
  oison_taxon \%>\%
  sf::st_as_sf() \%>\%
  add_type_geometrie()
}

}
