% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-connexion.R
\name{check_acces_oison_api}
\alias{check_acces_oison_api}
\title{Test si les accès à l'API OISON sont bien renseignés}
\usage{
check_acces_oison_api()
}
\value{
Aucune valeur. Une erreur est renvoyée si les accès ne sont pas renseignés.
}
\description{
Fonction pour tester si les accès login/mdp sont bien stockés dans
l'environnement R (fichier `.Renviron`).
}
\examples{
\dontrun{
check_acces_oison_api()
}
}
