% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dpt.R
\name{get_data_dpt}
\alias{get_data_dpt}
\title{Consulter et Télécharger les données OISON pour un département}
\usage{
get_data_dpt(dpt_code, login, mdp, collect_all = FALSE)
}
\arguments{
\item{dpt_code}{code INSEE de département (ex. 27)}

\item{login}{login du compte}

\item{mdp}{mot de passe associé au compte}

\item{collect_all}{Booléen. Si FALSE (par défaut) consultation du nombre d'observation (FALSE), si TRUE télécharge les données.}
}
\value{
Si collect_all=FALSE, retourne un dataframe avec le code du département et le nombre d'observation dans celui-ci.
Si collect_all=T, retourne un dataframe avec les observations.
}
\description{
Consulter et Télécharger les données OISON pour un département
}
\examples{
\dontrun{

# -> consulter le nombre de données dans le département 27

get_data_dpt(dpt_code = "27", login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp")

#' # -> télécharger les données dans le département 27

obs_dpt27 <-
get_data_dpt(dpt_code = "27", login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp", collect_all = T)
}

}
