% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_polygon.R
\name{get_data_polygon}
\alias{get_data_polygon}
\title{Consulter et Télécharger les données OISON sur une zone définie}
\usage{
get_data_polygon(
  zone_selection,
  login,
  mdp,
  collect_all = FALSE,
  draw_zone = FALSE,
  ...
)
}
\arguments{
\item{zone_selection}{un objet contenant le zone (polygon), RGF93 (epsg: 2154)}

\item{login}{login du compte}

\item{mdp}{mot de passe associé au compte}

\item{collect_all}{Booléen. Si FALSE (par défaut) consultation du nombre d'observation (FALSE), si TRUE télécharge les données.}

\item{draw_zone}{Booléen. Si TRUE, lance l'outil de sélection de zones. Par défaut, FALSE.}

\item{...}{}
}
\value{
Si collect_all=FALSE, retourne un dataframe avec le polygone et le nombre d'observation dans celui-ci.
Si collect_all=T, retourne un dataframe avec les observations.
}
\description{
Consulter et Télécharger les données OISON sur une zone définie
}
\examples{
\dontrun{
rm(selection_zone)

## -> collecte des donnees sur polygone defini manuellement

test_zone <-
"POLYGON ((410677.9 6877209, 413275.2 6929083, 496961.8 6925506, 495122.6 6873600, 410677.9 6877209))"

get_data_polygon(zone_selection = test_zone,
                 login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp",
                 collect_all = F, draw_zone = F)

## -> consulter des données sur polygone tracé en meme temps

get_data_polygon(login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp",
                 collect_all = F, draw_zone = T)

## -> collecte des données sur polygone tracé en meme temps

get_data_polygon(login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp",
                 collect_all = T, draw_zone = T)

## -> consulte des données sur polygon déjà dans l'environnement

get_data_polygon(login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp",
                 collect_all = F, draw_zone = F)

}

}
