% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_region.R
\name{get_data_region}
\alias{get_data_region}
\title{Consulter et Télécharger les données OISON pour une région}
\usage{
get_data_region(region_code, login, mdp, collect_all = FALSE)
}
\arguments{
\item{region_code}{code INSEE de la région (ex. 28 pour la Normandie)}

\item{login}{login du compte}

\item{mdp}{mot de passe associé au compte}

\item{collect_all}{Booléen. Si FALSE (par défaut) consultation du nombre d'observation (FALSE), si TRUE télécharge les données.}
}
\value{
Si collect_all=FALSE, retourne un dataframe avec le code de la région et le nombre d'observation dans celui-ci.
Si collect_all=T, retourne un dataframe avec les observations.
}
\description{
Consulter et Télécharger les données OISON pour une région
}
\examples{
\dontrun{

# -> consulter le nombre de données dans la région Île-de-France (code region 11)

get_data_region(region_code = "11", login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp")

#' # -> télécharger les données dans la région Île-de-France (code region 11)

obs_regIDF <-
get_data_region(region_code = "11", login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp", collect_all = T)
}

}
