% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_taxon_sql.R
\name{get_table_taxon_sql}
\alias{get_table_taxon_sql}
\title{Fonction pour créer le tableau taxon OISON via connexion SQL}
\usage{
get_table_taxon_sql(conn)
}
\arguments{
\item{conn}{nom de la base à utiliser.}
}
\value{
un dataframe avec les observations taxons
}
\description{
Permet de créer une 'table taxon' à partir de la base SQL.
La table ainsi créée correspond par défaut à l'ensemble des observations sur
les taxons bancarisés dans OISON.
}
\examples{
\dontrun{
bdd_oison <- start_sql_connexion()

df_taxon_oison_sql <- get_table_taxon_sql(conn = bd_oison)

stop_sql_connexion(conn = bdd_oison)
}

}
