% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon_dpt.R
\name{get_taxon_dpt}
\alias{get_taxon_dpt}
\title{Consulter et télécharger les données sur les taxons de la bdd OISON pour un département}
\usage{
get_taxon_dpt(
  dpt_code,
  login,
  mdp,
  collect_all = FALSE,
  date_min = NULL,
  date_max = NULL
)
}
\arguments{
\item{dpt_code}{code du département (ex. 27)}

\item{login}{login du compte OISON}

\item{mdp}{mot de passe associé au compte OISON}

\item{collect_all}{Booléen. Si FALSE (par défaut) consultation du nombre d'observation (FALSE).
Si TRUE, télécharge les données pour les taxons.}

\item{date_min}{date minimale pour la période au format (\code{\%Y-\%m-\%d} ; cf. exemple '2022-01-01').
NULL par défaut.}

\item{date_max}{date maximum pour la période au format (\code{\%Y-\%m-\%d} ; cf. exemple '2022-12-31').
NULL par défaut.}
}
\value{
Si \code{collect_all}=FALSE, retourne un dataframe indiquant le nombre d'observations dans le département.
 Si \code{collect_all}=TRUE, retourne un dataframe avec le détail des observations taxons.
 Si \code{date_min} et \code{date_max} sont spécifiés, retourne un dataframe avec les valeurs sur la période.
}
\description{
Consulter et télécharger les données sur les taxons de la bdd OISON pour un département
}
\examples{
\dontrun{
# -> Exemple pour consulter rapidement le nombre d'observations dans le département 27,
# sans téléchargement des données :
get_taxon_dpt(dpt_code = "27",
              login = "john.doe@ofb.gouv.fr",
              mdp = "mon_mdp")

# -> Exemple pour consulter rapidement le nombre d'observations sur la période 2022
# dans le département 27, sans téléchargement des données :
get_taxon_dpt(
  dpt_code = "27",
  login = "john.doe@ofb.gouv.fr",
  mdp = "mon_mdp",
  date_min = '2022-01-01',
  date_max = '2022-12-31'
)

# -> Exemple pour télécharger les données dans le département 27
obs_dpt27 <-
  get_taxon_dpt(login = "john.doe@ofb.gouv.fr",
                mdp = "mon_mdp",
                dpt_code = "27",
                collect_all = T)
}
}
