% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_importer_dump_oison.R
\name{imp_importer_dump_oison}
\alias{imp_importer_dump_oison}
\title{Fonction pour importer un dump de la base OISON}
\usage{
imp_importer_dump_oison(fichier_dump)
}
\arguments{
\item{fichier_dump}{chemin vers l'objet .sql du dump}
}
\value{
L'ensemble des tables de la base (sauf tables 'mapping')
}
\description{
Fonction pour importer un dump de la base OISON
}
\examples{

\dontrun{
imp_importer_dump_oison(fichier_dump = "./raw_data/PROD_oison.sql")
}

}
