% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_inpn_maille_taxon.R
\name{imp_inpn_maille_taxon}
\alias{imp_inpn_maille_taxon}
\title{Récupération des mailles 10x10km sur l'INPN sur une zone shape}
\usage{
imp_inpn_maille_taxon(df, shape)
}
\arguments{
\item{df}{un dataframe issu de la requete OISON}

\item{shape}{un dataframe contenant une couche shape pour la zone souhaitée}
}
\value{
un dataframe avec les mailles 10x10km INPN
}
\description{
Récupération des mailles 10x10km sur l'INPN sur une zone shape
}
\examples{
\dontrun{

data_oison <- get_data_dpt(dpt_code = "27", login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp")

normandie_shp <-
readRDS(file = './regions_geo_normandie.rds') \%>\%
sf::st_transform(crs = 4326)

maille_inpn <- imp_inpn_maille_taxon(df = data_oison,  shape = normandie_shp)

}

}
