% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_dump_creer_table_taxon.R
\name{mef_dump_creer_table_taxon}
\alias{mef_dump_creer_table_taxon}
\title{Fonction pour créer un tableau taxon issu d'un dump OISON}
\usage{
mef_dump_creer_table_taxon()
}
\value{
un dataframe avec les observations taxons.
 La colonne \code{geometry} spécifie le type de géométrie (point, lignes, polygones) pour l'observation,
 et doit être traitée avec précaution lors d'un export sous un autre format (.csv, ...).
}
\description{
Cette fonction permet de créer un tableau avec les observations \code{taxon} présentes dans la
 base OISON, à partir des tables importées par le dump de la base.
}
\details{
Nécessite au préalable d'importer le dump sql avec la fonction \code{imp_importer_dump_oison}.
}
\examples{
\dontrun{
df_taxon_oison <- mef_dump_creer_table_taxon()
}

}
