% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oisonR-package.R
\docType{package}
\name{oisonR-package}
\alias{oisonR}
\alias{oisonR-package}
\title{oisonR: R Package pour collecter et utiliser les données OISON}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Ensemble de fonctions pour utiliser les données bancarisées sur OISON, l'application pour les saisies opportunistes de l'Office français de la biodiversité (OFB).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richaben/oisonR}
  \item \url{https://richaben.github.io/oisonR/}
  \item Report bugs at \url{https://github.com/richaben/oisonR/issues}
}

}
\author{
\strong{Maintainer}: Benoit Richard \email{benoit.richard@ofb.gouv.fr} (\href{https://orcid.org/0000-0003-4522-027X}{ORCID})

}
\keyword{internal}
