% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outil_selection_zone.R
\name{outil_selection_zone}
\alias{outil_selection_zone}
\title{Outil de sélection de zones (polygones)}
\usage{
outil_selection_zone()
}
\value{
Lance une fenetre interactive Shiny pour selectionner son polygone et le sauver dans l'environnement pour une future requete
}
\description{
Outil de sélection de zones (polygones)
}
\examples{
\dontrun{
# -> Lancer l'outil et suivre les instructions
outil_selection_zone()
}
}
