% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-connexion.R
\name{stop_sql_connexion}
\alias{stop_sql_connexion}
\title{Stoppe une connexion à OISON via la base de données SQL}
\usage{
stop_sql_connexion(conn)
}
\arguments{
\item{conn}{nom de la base à stopper.}
}
\value{
Aucune valeur.
}
\description{
Fonction stoppe la connexion à la base.
}
\examples{
\dontrun{
bdd_oison <- start_sql_connexion()

stop_sql_connexion(conn = bdd_oison)
}

}
