% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_inpn_habitats_taxon.R
\name{imp_inpn_habitats_taxon}
\alias{imp_inpn_habitats_taxon}
\title{Récupération des codes Habitats des taxons sur l'INPN}
\usage{
imp_inpn_habitats_taxon(df)
}
\arguments{
\item{df}{un dataframe issu de la requete OISON}
}
\value{
un dataframe avec les codes habitats présents sur l'INPN
}
\description{
Récupération des codes Habitats des taxons sur l'INPN
}
\examples{
\dontrun{
data_oison <- get_data_dpt(dpt_code = "27",
login = "john.doe@ofb.gouv.fr", mdp = "mon_mdp")

code_habitats <- imp_inpn_habitats_taxon(df = data_oison)

}
}
