% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outil_selection_zone.R
\name{outil_selection_zone}
\alias{outil_selection_zone}
\title{Outil interactif pour tracer une zone (type polygone)}
\usage{
outil_selection_zone()
}
\value{
Lance une fenêtre interactive Shiny pour sélectionner son polygone et
 le sauver dans l'environnement pour une future requête
 avec la fonction \code{get_taxon_polygon}.
}
\description{
Outil interactif pour tracer une zone (type polygone)
}
\examples{
\dontrun{
# -> Lancer l'outil et suivre les instructions
outil_selection_zone()
}
}
